[bugfix] Move methods to create models to `Mdb`
